Updated checker to use XML parsing
Opening the file in text mode is not necessarily possible as the
utf8 file may content text segments encoded differently. It is
unclear whether this is allowed by the ImzML spec but this solves
the issue.

This commit also fixes an error where the optional braces would be
included in the UUID from the .imzml file